Remove mermaid (#41)
Initially we had bundled mermaid into easy_infra because at Seiso we
feel strongly that documentation should be tightly coupled with your
code and/or configuration, and when possible, be automatically
generated. However, this caused significant image bloat and we decided
to remove it for the time being, and in the future we may release a
separate documentation generation image to fulfill this use case.