Remove caching
The speed improvement isn't worth the added complexity of both
implementing appropriate caching, as well as accounting for potential
issues in CI caused by cache hits. By removing this we can ensure we
always have a fresh start, and at the end of the day the speed
improvements have not been substantial.